Add homepage URL to package.json
Setting the homepage URL in package.json to
"http://ali-muhammed-dev.tech/Ahel-El-Quran/" for the project. This
change is made to provide a public URL for accessing the Quran project.
This URL can be used for showcasing, sharing, and accessing the project
online. No issues are associated with these changes.

Related: None