Add redux-thunk middleware for asynchronous actions
The redux-thunk middleware is added to the store configuration in order
to handle asynchronous actions in the Redux application. This enables
the use of the fetchLocation action, which dispatches an API request to
get the user's location based on their coordinates. The received
location is stored in localStorage and used to display the city and
country in the Location component. The addition of redux-thunk allows
for better management of asynchronous logic in the Redux application.

Closes #1234