Refactor Azkar component to toggle display of Azkar names
This commit refactors the Azkar component to incorporate a feature that
allows the user to toggle the display of Azkar names. The addition of
the `showAzkarNames` state variable, along with the corresponding click
event handling, enables the user to show or hide the names of the Azkar.

The changes include:
- Adding the `showAzkarNames` state variable using the `useState` hook
- Implementing the `handleOpenButtonClick` function to toggle the value
of `showAzkarNames`
- Modifying the JSX to conditionally render the Azkar names based on the
state of `showAzkarNames`

These changes aim to enhance the user experience and provide more
flexibility in navigating and interacting with the Azkar component. No
issues are associated with this commit.

Note: CSS changes related to responsiveness are not highlighted in this
commit message.