Improve Google Sign-In button UI
The UI of the Google Sign-In button has been enhanced by adding a CSS
class to provide better styling. The addition of the "m-5" class ensures
proper spacing between the button and surrounding elements. This change
improves the visual appeal and user experience of the login process. No
relevant issue references.

Note: The commit message assumes that the "m-5" class has been
previously defined in the CSS file or stylesheet.